+++ src/Controllers/API/UserAccountsController.cs
@@ -12,51 +12,82 @@ namespace TYRISKANALYSIS.Controllers.API
     public class UserAccountController : Controller
     {
         private IUserAccountRepository repository;
+        private IUserAccountRepository authenticate;
 
         public UserAccountController(IConfiguration appconfig)
         {
             repository = new UserAccountRepository(appconfig);
+            authenticate = new UserAccountRepository(appconfig);
         }
 
         // GET api/useraccounts
         [HttpGet]
         public IActionResult GetUserAccount(PagedParams param)
         {
-            var request = this.Request.QueryString;
-            return Ok(repository.GetAll(param));
+            var token = this.Request.Headers["tokenAuthorization"];
+            if (authenticate.GetByToken(token))
+            {
+                var request = this.Request.QueryString;
+                return Ok(repository.GetAll(param));
+            }
+            else
+            {
+                return BadRequest();
+            }
         }
 
         // GET api/useraccounts/1
         [HttpGet("{id}")]
         public IActionResult GetUserAccount(int id)
         {
-            var userAccount = repository.GetById(id);
-
-            if (userAccount == null)
+            var token = this.Request.Headers["tokenAuthorization"];
+            if (authenticate.GetByToken(token))
             {
-                return NotFound();
-            }
+                var userAccount = repository.GetById(id);
 
-            return Ok(userAccount);
+                if (userAccount == null)
+                {
+                    return NotFound();
+                }
+
+                return Ok(userAccount);
+            } else
+            {
+                return BadRequest();
+            }
         }
 
         // POST api/useraccounts
         [HttpPost]
         public IActionResult SaveUserAccount([FromBody] UserAccount userAccount)
         {
-            if (userAccount.Id == 0)
-                repository.Add(userAccount);
-            else
-                repository.Update(userAccount);
-            return Ok(userAccount);
+            var token = this.Request.Headers["tokenAuthorization"];
+            if (authenticate.GetByToken(token))
+            {
+                if (userAccount.Id == 0)
+                    repository.Add(userAccount);
+                else
+                    repository.Update(userAccount);
+                return Ok(userAccount);
+            } else
+            {
+                return BadRequest();
+            }
         }
 
         // DELETE api/useraccounts/1
         [HttpDelete("{id}")]
         public IActionResult DeleteUserAccount(int id)
         {
-            repository.Remove(id);
-            return Ok();
+            var token = this.Request.Headers["tokenAuthorization"];
+            if (authenticate.GetByToken(token))
+            {
+                repository.Remove(id);
+                return Ok();
+            } else
+            {
+                return BadRequest();
+            }
         }
 
         // POST api/useraccounts
